+++ CHANGELOG.md
@@ -39,6 +39,12 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
   [#606](https://github.com/hynek/structlog/pull/606)
 
 
+### Fixed
+
+- `structlog.stdlib.render_to_log_kwargs` now correctly passes stacklevel as a kwarg to stdlib logging.
+  [#619](https://github.com/hynek/structlog/issues/619)
+
+
 
 ## [24.1.0](https://github.com/hynek/structlog/compare/23.3.0...24.1.0) - 2024-01-08